Updated agents.py to the latest:
Moved other agents.py file to agents-bak.py.

Removed all extraneous code (Redundant Agents/Objects, 1D world, Wumpus World, etc.)

Began the process of pulling in useful code from the agents-bak.py file into agents.py.
